requirements.txt
tensorflow
pandas
numpy
dvc
notebook
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-pyYAML
scipy
Flask
Flask-Cors
kaggle
subprocess

-e .


